Use path ID to filter in-flight counter decrements
Identifying the path associated with a sent packet based on packet
number alone is only reliable if the path is still known, but we may
forget paths quickly e.g. when path validation fails or is abandoned.

We could alternatively retain information about abandoned paths so
long as they have in-flight packets, but this is simpler.